Bundle own manifest, instead of using shell32's
We write the manifest to a temporary file. We then use the temporary file, as opposed to shell32, to create the activation context. This continues to avoid the need to link in the manifest as a resource.

Bundling our own manifest should be more future proof than using shell32's. In fact, the approach now is analogous to what WinForms does, which bodes well for compatibility.

Fixes #218.